Refactor tests
- Move tests to tests/routes/
- Use a the Sinon sandbox object for easy test cleanup
