+++ tests/.eslintrc
@@ -1,8 +1,14 @@
-# Include a .eslintrc for the tests directory, unless someone
-# is willing to refactor and lint the rest of the codebase.
+# Include a .eslintrc for the tests directory.
+#
 # Mocha discourages arrow callbacks. Override some rules to allow regular function callbacks.
+# This allows us to use the `this` context to share variables between `beforeEach()` and `it()` blocks.
+#
+# The alternative is to use `let user;` in the root of the `describe()` block and assign
+# `user = 'something';` in inner blocks. This works well, but I find it tends to get messy.
+#
+# Having a `this.user` prefix is more obvious when scanning code.
+# Thus the decision to stick with `function()` callbacks is simply choosing the lesser evil.
 {
-  "extends": "airbnb",
   "rules": {
     "prefer-arrow-callback": 0,
     "func-names": 0,
